src/main/java/me/dolia/pmm/role/RoleService.java
package me.dolia.pmm.role;

import me.dolia.pmm.service.NotFoundException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Service to manage {@link Role} instances.
 */
@Service
public class RoleService {

    private final RoleRepository roleRepository;

    @Autowired
    public RoleService(RoleRepository roleRepository) {
        this.roleRepository = roleRepository;
    }

    public Role find(String name) {
        return roleRepository.findByName(name).orElseThrow(() -> new NotFoundException(Role.class, "name", name));
    }

    public Role save(Role role) {
        return roleRepository.save(role);
    }

    public void delete(Role role) {
        roleRepository.delete(role);
    }
}